+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/Fields/index.js
@@ -1,0 +1,8 @@
+/**
+ * Generic input fields.
+ */
+
+export { default as PayPalCheckbox } from './Checkbox';
+export { default as CheckboxGroup } from './CheckboxGroup';
+export { default as PayPalRdb } from './RadioButton';
+export { default as PayPalRdbWithContent } from './RadioContent';
